first cut, moves fxa creds and prompts
