+++ components/logins/Cargo.toml
@@ -30,6 +30,6 @@ more-asserts = "0.2.1"
 env_logger = "0.5.13"
 prettytable-rs = "0.7.0"
 fxa-client = { path = "../fxa-client" }
-webbrowser = "0.3.1"
 chrono = "0.4.6"
 clap = "2.32.0"
+cli-support = { path = "../support/cli" }
